requirements.txt
setuptools
psutil
python-telegram-bot
PySocks
